simStep now uses double times as well, refs #912

+++ src/traci_testclient/TraCITestClient.h
@@ -69,7 +69,7 @@ protected:
     /** @brief Sends and validates a simulation step command
      * @param[in] time The time step to send
      */
-    void commandSimulationStep(int time);
+    void commandSimulationStep(double time);
 
 
     /** @brief Sends and validates a Close command
